+++ src/module/actor/sheets/charactersheet-v2.ts
@@ -1,3 +1,4 @@
+import { ISCharacterTour } from '../../features/tours/is-character-tour'
 import { IronswornSettings } from '../../helpers/settings'
 import characterSheetVue from '../../vue/character-sheet.vue'
 import { VueActorSheet } from '../../vue/vueactorsheet'
@@ -13,12 +14,18 @@ export class IronswornCharacterSheetV2 extends VueActorSheet {
     }) as any
   }
 
+  activateTab(tabKey: string) {
+    this.localEmitter.emit('activateTab', tabKey)
+  }
+
   render(...args: any[]) {
     if (this._state <= Application.RENDER_STATES.NONE) this._openMoveSheet()
     return super.render(...args) as any
   }
 
   _getHeaderButtons() {
+    const [editButton, sheetButton, tokenButton, ...otherButtons] =
+      super._getHeaderButtons()
     return [
       {
         class: 'ironsworn-open-move-sheet',
@@ -26,7 +33,16 @@ export class IronswornCharacterSheetV2 extends VueActorSheet {
         icon: 'fas fa-directions',
         onclick: (e) => this._openMoveSheet(e),
       },
-      ...super._getHeaderButtons(),
+      editButton,
+      sheetButton,
+      tokenButton,
+      {
+        class: 'ironsworn-help',
+        icon: 'fa fa-circle-question',
+        label: '',
+        onclick: (e) => new ISCharacterTour(this.actor).start(),
+      },
+      ...otherButtons,
     ]
   }